fix(reborn): address event stream review follow-ups
