+++ crates/ironclaw_event_streams/src/update_source.rs
@@ -1,0 +1,56 @@
+use async_trait::async_trait;
+use ironclaw_event_projections::ProjectionScope;
+use ironclaw_turns::TurnActor;
+use serde::{Deserialize, Serialize};
+use tokio::sync::broadcast;
+
+use crate::{
+    error::ProjectionStreamError,
+    types::{ProductProjectionEnvelope, ProjectionTarget, ProjectionViewClass},
+};
+
+#[async_trait]
+pub trait ProjectionUpdateSource: Send + Sync {
+    async fn subscribe(
+        &self,
+        request: ProjectionLiveUpdateRequest,
+    ) -> Result<broadcast::Receiver<ProductProjectionEnvelope>, ProjectionStreamError>;
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
+pub struct ProjectionLiveUpdateRequest {
+    pub actor: TurnActor,
+    pub scope: ProjectionScope,
+    pub view: ProjectionViewClass,
+    pub target: ProjectionTarget,
+}
+
+pub struct InMemoryProjectionUpdateSource {
+    sender: broadcast::Sender<ProductProjectionEnvelope>,
+}
+
+impl InMemoryProjectionUpdateSource {
+    pub fn new(capacity: usize) -> Self {
+        let (sender, _) = broadcast::channel(capacity.max(1));
+        Self { sender }
+    }
+
+    pub fn publish(
+        &self,
+        envelope: ProductProjectionEnvelope,
+    ) -> Result<usize, ProjectionStreamError> {
+        self.sender
+            .send(envelope)
+            .map_err(|_| ProjectionStreamError::Source)
+    }
+}
+
+#[async_trait]
+impl ProjectionUpdateSource for InMemoryProjectionUpdateSource {
+    async fn subscribe(
+        &self,
+        _request: ProjectionLiveUpdateRequest,
+    ) -> Result<broadcast::Receiver<ProductProjectionEnvelope>, ProjectionStreamError> {
+        Ok(self.sender.subscribe())
+    }
+}
